+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/RequestParser.kt
@@ -19,7 +19,7 @@ import io.ktor.util.*
 import kotlinx.serialization.PolymorphicSerializer
 import kotlinx.serialization.json.JsonObject
 
-internal class RequestParser(
+class RequestParser(
     /**
      * Need to implement if resources are defined by reference, i.e. the URL for a [at.asitplus.signum.indispensable.josef.JsonWebKeySet],
      * or the request itself as `request_uri`, or `presentation_definition_uri`.